pyproject.toml
[tool.poetry]
name = "ai-voice-chat"
version = "0.1.0"
description = ""
authors = ["yqq <youngqqcn@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
openai = "^1.54.3"
dashscope = "^1.20.12"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
